Split unit tests from feature tests
